pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "daily-learner"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = ["click", "rich", "pyyaml", "requests"]

[project.scripts]
learner = "src.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]
